+++ zcash_keys/CHANGELOG.md
@@ -11,7 +11,6 @@ and this library adheres to Rust's notion of
 - `zcash_keys::address::Address::has_receiver`
 - `impl Display for zcash_keys::keys::AddressGenerationError`
 - `impl std::error::Error for zcash_keys::keys::AddressGenerationError`
-- `zcash_keys::keys::UnifiedError`
 - `zcash_keys::keys::UnifiedFullViewingKey::from_ufvk`
 - `zcash_keys::keys::UnifiedFullViewingKey::to_ufvk`
 - `zcash_keys::keys::UnifiedFullViewingKey::to_unified_incoming_viewing_key`
@@ -25,8 +24,8 @@ and this library adheres to Rust's notion of
   (instead of `Option<(UnifiedAddress, DiversifierIndex)>` for `find_address`).
 - `zcash_keys::keys::AddressGenerationError`
   - Dropped `Clone` trait
-  - Added `UnifiedError` variant.
-  - Added `HDWalletError` variant.
+- `zcash_keys::keys::DerivationError`
+  - Added `InvalidShieldedKey` variant.
 
 ### Fixed
 - `UnifiedFullViewingKey::find_address` can now find an address for a diversifier